03_WEB_APPLICATIONS/STAGE_3_APP/BE/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv